+++ mod/src/main/java/basemod/BaseMod.java
@@ -206,6 +206,7 @@ public class BaseMod {
 	private static HashMap<String, String> keywordProperNames;
 	private static HashMap<String, String> keywordUniqueNames;
 	private static HashMap<String, String> keywordUniquePrefixes;
+	private static HashMap<String, Color> keywordColors;
 
 	public static ArrayList<String> encounterList;
 	public static HashMap<String, String> underScoreEncounterIDs;
@@ -396,6 +397,7 @@ public class BaseMod {
 		keywordProperNames = new HashMap<>();
 		keywordUniqueNames = new HashMap<>();
 		keywordUniquePrefixes = new HashMap<>();
+		keywordColors = new HashMap<>();
 		BaseModInit baseModInit = new BaseModInit();
 		BaseMod.subscribe(baseModInit);
 
@@ -1562,6 +1564,10 @@ public class BaseMod {
 	}
 
 	public static void addKeyword(String modID, String proper, String[] names, String description) {
+		addKeyword(modID, proper, names, description, null);
+	}
+
+	public static void addKeyword(String modID, String proper, String[] names, String description, Color color) {
 		if (modID != null && !modID.isEmpty()) {
 			if (!modID.endsWith(":")) {
 				modID = modID + ":";
@@ -1576,6 +1582,7 @@ public class BaseMod {
 			}
 		}
 		String parent = names[0];
+		keywordColors.put(parent, color);
 
 		if (proper != null) {
 			keywordProperNames.put(parent, proper);
@@ -1602,6 +1609,10 @@ public class BaseMod {
 
 	public static String getKeywordPrefix(String keyword) {
 		return keywordUniquePrefixes.get(keyword);
+	}
+
+	public static Color getKeywordColor(String keyword) {
+		return keywordColors.get(keyword);
 	}
 
 	public static String getKeywordTitle(String keyword) {